Refactor create watches
